Fixed a big and added blank tile sprite
Fixed the bug with extra inputs going through after questions.

Added the blank tile sprite, so after answering a question, it places a blank tile on the current tile.

Also changed the spawntile function to instead of create GameObjects of the sprites, change the sprites of GameObjects declared once at the start.